refactor(repos): extract /api/stats and /api/usage into repos_stats.py
`backend/routers/repos.py` was 505 lines, exceeding the 500-line soft
cap enforced by `ci-standard.yml`. Split the stats and usage-monitoring
routes into a sibling `repos_stats.py` module and re-include its router
from `repos.router`, so callers register a single router as before.
Dependency injection still flows through `repos.set_dependencies`.

No behavior change.
